Connect add book and add reader window

+++ src/main/java/pl/wietwioorki/to22019/model/Reader.java
@@ -10,7 +10,7 @@ import java.util.Date;
 @Getter
 @ToString
 public class Reader {
-    private int pesel;
+    private long pesel;
     private String fullName;
     private Date dateOfBirth;
 }
